djangorest/requirements.txt
django==3.1.4
django-rest-framework==0.1.0
coreapi===2.3.3
django-filter==2.3.0
django-environ==0.4.5
django-cors-headers==3.4.0
pyjwt==1.7.1
drf-jwt==1.19.1
requests==2.24.0
boto3==1.16.47


